fix: page switcher get the wrong pages status
