+++ .changeset/seven-taxis-happen.md
@@ -1,0 +1,6 @@
+---
+'@sveltepress/theme-default': patch
+'@sveltepress/create': patch
+---
+
+fix: page switcher get the wrong pages status
