+++ packages/theme-default/src/components/Toc.svelte
@@ -78,7 +78,7 @@
 
 <style>
   .toc {
-    --at-apply: 'transition-transform transition transition-500 py-4 text-gray-5 dark:text-gray-2 sm:z-3 leading-[2em] bottom-0 right-0 sm:top-[80px] fixed text-3.5 sm:w-[22vw] w-[60vw] bg-white dark:bg-zinc-8 sm:bg-transparent top-0 z-999 sm:dark:bg-transparent';
+    --at-apply: 'transition-transform transition transition-300 py-4 text-gray-5 dark:text-gray-2 sm:z-3 leading-[2em] bottom-0 right-0 sm:top-[80px] fixed text-3.5 sm:w-[22vw] w-[60vw] bg-white dark:bg-zinc-8 sm:bg-transparent top-0 z-999 sm:dark:bg-transparent';
   }
   .title {
     --at-apply: 'font-bold pl-4 text-gray-8 dark:text-gray-2';
